Fix issue with duplicate scoped css folders during watch
Fixes an issue where the sass process fails to start when the "ScopedCssFolders" appsettings is used with some of the default values.

It will now initialize the options without ScopedCssFolders, and when options are validated it will set it to the default folders if the value is null.
Also guard against a user specifying the same directory multiple times by using a HashSet with dictionaries instead of List