fix deadlocking when reading StandardOutput and StandardError
When processing input that generates enough output to StandardError to fill the buffer, the current implementation is awaiting reading from StandardOutput whilst the child process is waiting for its now full StandardError buffer to be read from in order to continue processing.
This change now sees StandardOutput and StandardError being read from asynchronously at the same time which prevents the buffers from filling up and causing a deadlock.